Synchronized Packet Delivery
In CentOS, seeing buffering of data in multiples of 1460 Bytes. The following is a quick iterative change to loop over the pending message of "bufferSize" from a client, and similarly to send "bufferSize" length of an echo back to client. Next iteration will account for variable message size from client; rather than strictly expecting "bufferSize" length.