Synchronized Variable Recv Size
OS handling the streaming buffer of TCP connection means going to an intermediary buffer before recv collects the message. Because of this buffering, entire message length not necessarily received, and so polling until the end of the message size becomes necessary. 1 GB RAM CentOS server seems to use a buffer size of 1460 Bytes - so only multiples of 1460B received at a time on (this) server.

Server echoes back the message by keeping a sum of the total Bytes received in the iterative "receivingMessage" function. Reading various manuals, unable to determine if this is the case with sending. Both Windows and CentOS server seem happy to "send" full message size from one invocation, but similar "sendingMessage" function provided in case it is chunked by an OS managed streaming buffer.

Note, this should be different than UDP which explicitly sends as packets, not as an open stream...