Tidying TCP Server Program
Comments which would might hurt the transmission time of a round trip message have been removed. Buffer chunk free function to extend the "send" function has been commented out. Storing the raw bytes of the time data in a binary file has been removed (unnecessary for general use). Comment added to final manipulation of "result" integer to explain pre-processing of time data sent by client. Comments added to distinguish converting time data to uint32_t, and to write to a file - including alternative use case for converting during the same iterations as writing to a file. Commented out "fcntl" invocation with explanation that it was used to test for blocking or non-blocking with the receive function calls.