Prior Performance Client UDP
This is the original performance benchmark for the HL2 client application when using UDP. It might need to be updated when communicating with a remote server to ensure all packet deliveries are gathered before the server sends them back (similar to the issue faced when making the TCP stream remote). Note, the maximum buffer size is substantially smaller than the TCP version - but this quantity was found when testing on Windows LAN, so it might need to be modified.